fix(upstream): widen connection usage counters to u32
- use AtomicU32 for H2, H3, and DoQ connection usage counts
- widen Connection::using_count() to u32
- preserve UDP/TCP request-map counts with lossless conversion
- add regression coverage past u16::MAX

+++ src/infra/network/upstream/conn/tcp.rs
@@ -166,8 +166,8 @@ impl Connection for TcpConnection {
         }
     }
 
-    fn using_count(&self) -> u16 {
-        self.request_map.size()
+    fn using_count(&self) -> u32 {
+        u32::from(self.request_map.size())
     }
 
     fn available(&self) -> bool {
